third_party/blink/web_tests/platform/generic/fast/forms/number/number-keyoperation-expected.txt

Test for keyboard operations for <input type=number>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Inserting "ab123cd":
PASS input.value is "123"
Press the up arrow key:
PASS input.value is "124"
Press the down arrow key:
PASS input.value is "123"
Press the down and alt arrow key, should not decrement value:
PASS input.value is "123"
Disable input element:
PASS input.value is "123"
Read-only input element:
PASS input.value is "123"
PASS successfullyParsed is true

TEST COMPLETE
